# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Portfolio Platform API",
        "description": "Backend API for the AI-enabled dynamic portfolio platform",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:9100",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "List published projects",
                "description": "Ordered by display order. Not paginated. Optional ?search= matches title, description and technology names (FR-18).",
                "operationId": "listPublished_1",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListProjectSummaryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/media/{id}/content": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Serve a file",
                "description": "Public: the Public Site renders these to anonymous visitors.",
                "operationId": "content",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MediaResponse": {
                "type": "object",
                "description": "An uploaded file",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "url": {
                        "type": "string",
                        "description": "API path that serves the bytes",
                        "example": "/api/v1/media/1/content"
                    },
                    "fileName": {
                        "type": "string"
                    },
                    "originalFileName": {
                        "type": "string"
                    },
                    "mimeType": {
                        "type": "string"
                    },
                    "sizeBytes": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "storageBackend": {
                        "type": "string",
                        "enum": [
                            "LOCAL",
                            "OBJECT_STORAGE"
                        ]
                    },
                    "width": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "height": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "altText": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ApiResponseListProjectSummaryResponse": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProjectSummaryResponse"
                        }
                    },
                    "message": {
                        "type": "string"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ProjectSummaryResponse": {
                "type": "object",
                "description": "A project as it appears in a list",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "shortDescription": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "$ref": "#/components/schemas/MediaResponse"
                    },
                    "technologies": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "featured": {
                        "type": "boolean"
                    },
                    "projectType": {
                        "type": "string",
                        "enum": [
                            "PERSONAL",
                            "PROFESSIONAL",
                            "OPEN_SOURCE",
                            "CLIENT"
                        ]
                    },
                    "startDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "endDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "displayOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "PUBLISHED",
                            "ARCHIVED"
                        ]
                    }
                }
            },
...
        }
    }
}